requirements.txt
pyg-lib==0.1.0
torch-scatter==2.1.0
torch-sparse==0.6.16
torch-cluster==1.6.0
torch-spline-conv==1.2.1
torch-geometric==2.2.0
pandas
matplotlib
gym
kubernetes